Added normalization for cost functions.
Normalization is not used as it slows down the computations a lot. Too many calls to externals functions.
Weights are normalized using a diagonal matrix, but still is not effective enough as they are not correctly tuned.
First idea was to use the MAximum and minimum value produced by the single objective cases.